refactor: rename MapLibreRN.ts
